requirements.txt
Flask==2.3.3
supabase==1.0.3
python-dotenv==1.0.0
qrcode[pil]==7.4.2
pytz==2023.3
Pillow==10.0.1
